tests: Add more D3D11 tests Move mocks into separate file.
